# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Inspector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InspectInput {
  /** The URL to inspect */
  url: string;
}

export type TechDetectionConfidence = typeof TechDetectionConfidence[keyof typeof TechDetectionConfidence];


export const TechDetectionConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface TechDetection {
  /** Technology name (e.g. React, WordPress, nginx) */
  name: string;
  /** Category (e.g. Framework, CMS, Server, Library, Analytics) */
  category: string;
  /**
     * Detected version if available
     * @nullable
     */
  version?: string | null;
  confidence: TechDetectionConfidence;
  /** What triggered the detection (e.g. header X-Powered-By, script src pattern) */
  evidence: string;
}

export interface HeaderEntry {
  name: string;
  value: string;
}

export interface MetaTag {
  name: string;
  content: string;
}

export interface AssetLink {
  url: string;
  type: string;
  /** @nullable */
  size?: number | null;
}

export interface PageLink {
  href: string;
  text: string;
  isInternal: boolean;
}

export interface FormField {
  name: string;
  type: string;
}

export interface FormInfo {
  action: string;
  method: string;
  fields: FormField[];
}

export interface InspectResult {
  url: string;
  /** URL after redirects */
  finalUrl: string;
  statusCode: number;
  /** Raw HTML of the page */
  html: string;
  title: string;
  /** Detected technologies, frameworks, CMS, and libraries */
  technologies: TechDetection[];
  responseHeaders: HeaderEntry[];
  metaTags: MetaTag[];
  cssLinks: AssetLink[];
  jsScripts: AssetLink[];
  images: AssetLink[];
  links: PageLink[];
  forms: FormInfo[];
  cookies: string[];
  wordCount: number;
  byteSize: number;
  inspectedAt: string;
}

export interface ResourceInput {
  url: string;
}

export interface ResourceResult {
  url: string;
  content: string;
  contentType: string;
}

export interface ErrorResponse {
  error: string;
}