# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Web Inspector API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: inspector
    description: Web inspection operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /inspect:
    post:
      operationId: inspectUrl
      tags: [inspector]
      summary: Inspect a URL
      description: Fetches a webpage and returns its HTML, CSS, JS, links, meta tags, headers, and asset inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectInput"
      responses:
        "200":
          description: Inspection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Fetch error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /fetch-resource:
    post:
      operationId: fetchResource
      tags: [inspector]
      summary: Fetch a single resource (CSS or JS)
      description: Fetches the raw content of a CSS or JS file given its URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceInput"
      responses:
        "200":
          description: Resource content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Fetch error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    InspectInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: The URL to inspect

    InspectResult:
      type: object
      required: [url, finalUrl, statusCode, html, title, responseHeaders, metaTags, cssLinks, jsScripts, images, links, forms, cookies, wordCount, byteSize, inspectedAt, technologies]
      properties:
        url:
          type: string
        finalUrl:
          type: string
          description: URL after redirects
        statusCode:
          type: integer
        html:
          type: string
          description: Raw HTML of the page
        title:
          type: string
        technologies:
          type: array
          items:
            $ref: "#/components/schemas/TechDetection"
          description: Detected technologies, frameworks, CMS, and libraries
        responseHeaders:
          type: array
          items:
            $ref: "#/components/schemas/HeaderEntry"
        metaTags:
          type: array
          items:
            $ref: "#/components/schemas/MetaTag"
        cssLinks:
          type: array
          items:
            $ref: "#/components/schemas/AssetLink"
        jsScripts:
          type: array
          items:
            $ref: "#/components/schemas/AssetLink"
        images:
          type: array
          items:
            $ref: "#/components/schemas/AssetLink"
        links:
          type: array
          items:
            $ref: "#/components/schemas/PageLink"
        forms:
          type: array
          items:
            $ref: "#/components/schemas/FormInfo"
        cookies:
          type: array
          items:
            type: string
        wordCount:
          type: integer
        byteSize:
          type: integer
        inspectedAt:
          type: string

    HeaderEntry:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string

    MetaTag:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        content:
          type: string

    AssetLink:
      type: object
      required: [url, type]
      properties:
        url:
          type: string
        type:
          type: string
        size:
          type: ["integer", "null"]

    PageLink:
      type: object
      required: [href, text, isInternal]
      properties:
        href:
          type: string
        text:
          type: string
        isInternal:
          type: boolean

    FormInfo:
      type: object
      required: [action, method, fields]
      properties:
        action:
          type: string
        method:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FormField"

    FormField:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string

    ResourceInput:
      type: object
      required: [url]
      properties:
        url:
          type: string

    ResourceResult:
      type: object
      required: [url, content, contentType]
      properties:
        url:
          type: string
        content:
          type: string
        contentType:
          type: string

    TechDetection:
      type: object
      required: [name, category, confidence, evidence]
      properties:
        name:
          type: string
          description: Technology name (e.g. React, WordPress, nginx)
        category:
          type: string
          description: Category (e.g. Framework, CMS, Server, Library, Analytics)
        version:
          type: ["string", "null"]
          description: Detected version if available
        confidence:
          type: string
          enum: [high, medium, low]
        evidence:
          type: string
          description: What triggered the detection (e.g. header X-Powered-By, script src pattern)

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string