# lib/api-zod/src/generated/types/inspectResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Inspector API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetLink } from './assetLink';
import type { FormInfo } from './formInfo';
import type { HeaderEntry } from './headerEntry';
import type { MetaTag } from './metaTag';
import type { PageLink } from './pageLink';
import type { TechDetection } from './techDetection';

export interface InspectResult {
  url: string;
  /** URL after redirects */
  finalUrl: string;
  statusCode: number;
  /** Raw HTML of the page */
  html: string;
  title: string;
  /** Detected technologies, frameworks, CMS, and libraries */
  technologies: TechDetection[];
  responseHeaders: HeaderEntry[];
  metaTags: MetaTag[];
  cssLinks: AssetLink[];
  jsScripts: AssetLink[];
  images: AssetLink[];
  links: PageLink[];
  forms: FormInfo[];
  cookies: string[];
  wordCount: number;
  byteSize: number;
  inspectedAt: string;
}