# lib/api-zod/src/generated/types/techDetection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Inspector API
 * OpenAPI spec version: 0.1.0
 */
import type { TechDetectionConfidence } from './techDetectionConfidence';

export interface TechDetection {
  /** Technology name (e.g. React, WordPress, nginx) */
  name: string;
  /** Category (e.g. Framework, CMS, Server, Library, Analytics) */
  category: string;
  /**
     * Detected version if available
     * @nullable
     */
  version?: string | null;
  confidence: TechDetectionConfidence;
  /** What triggered the detection (e.g. header X-Powered-By, script src pattern) */
  evidence: string;
}